Version Fonctionnelle

+++ src/Servo.cpp
@@ -16,39 +16,39 @@ Servo::Servo(ModuleType mt, BusCommunication busCom,
 Servo::~Servo() {
 }
 
-void Servo::processOutput() {
-	if (!1/*sendOutput*/) {
-		boolean new_val = false;
-		for (int iiii = 0; iiii < 5; iiii++) {
-			if (servoAngles[iiii] != servoTargetAngles[iiii]) {
-				new_val = true;
-			}
-		}
-
-		if (new_val) { // || (millis() >= lastWirting + 1000)) {
-			if (millis() >= lastReading + 10) {
-				Serial.print(StrIndent);
-				Serial.print("Sout:");
-				aJsonObject* objectJSON = aJson.createObject();
-				aJson.addItemToObject(objectJSON, "sn",
-						aJson.createItem((int) getSerialNumber()));
-				char buf[2];
-				for (int iiii = 0; iiii < 5; iiii++) {
-					if (servoAngles[iiii] != servoTargetAngles[iiii]) {
-						itoa(iiii + 1, buf, 10);
-						aJson.addItemToObject(objectJSON, buf,
-								aJson.createItem(servoTargetAngles[iiii]));
-
-						servoAngles[iiii] = servoTargetAngles[iiii];
-					}
-				}
-
-				sendOutput(objectJSON);
-				aJson.deleteItem(objectJSON);
-				lastWriting = millis();
-			}
+void Servo::sendJson() {
+	boolean new_val = false;
+	for (int iiii = 0; iiii < 5; iiii++) {
+		if (servoAngles[iiii] != servoTargetAngles[iiii]) {
+			new_val = true;
 		}
 	}
+
+	if (new_val && millis() >= lastWriting + 10) {
+
+		aJsonObject* objectJSON = toJson();
+		Serial.print(StrIndent);
+		Serial.print("Sout:");
+		sendOutput(objectJSON);
+		aJson.deleteItem(objectJSON);
+		lastWriting = millis();
+	}
+}
+aJsonObject* Servo::toJson() {
+	aJsonObject* objectJSON = aJson.createObject();
+	aJson.addItemToObject(objectJSON, "sn",
+			aJson.createItem((int) getSerialNumber()));
+	char buf[2];
+	for (int iiii = 0; iiii < 5; iiii++) {
+		if (servoAngles[iiii] != servoTargetAngles[iiii]) {
+			itoa(iiii + 1, buf, 10);
+			aJson.addItemToObject(objectJSON, buf,
+					aJson.createItem(servoTargetAngles[iiii]));
+
+			servoAngles[iiii] = servoTargetAngles[iiii];
+		}
+	}
+	return objectJSON;
 }
 
 void Servo::setServoAngle(uint8_t id, uint8_t angle) {
